Switch to using decimal
refinement of what IsDimensionless is

+++ NMeasure.Tests/StandardConfigurationTests.cs
@@ -14,22 +14,22 @@ namespace NMeasure.Tests
 
         public IEnumerable<object[]> BasicConversionChecksSource()
         {
-            yield return new object[] { U.Centimeter, U.Meter, 1, 0.01 };
-            yield return new object[] { U.Mile, U.Inch, 1, 63360 };
-            yield return new object[] { U.Day, U.Second, 1, 86400 };
-            yield return new object[] { U.Kilogram, U.Ounce, 1, 35.2740 };
-            yield return new object[] { U.Fahrenheit, U.Celsius, 100, 37.7778 };
-            yield return new object[] { U.Fahrenheit, U.Kelvin, 100, 310.9278 };
-            yield return new object[] { U.Pascal, U.Psi, 1, 6894 };
-            yield return new object[] { U.Kilometer, U.Mile, 100, 62.1371 };
+            yield return new object[] { U.Centimeter, U.Meter, 1m, 0.01m };
+            yield return new object[] { U.Mile, U.Inch, 1m, 63360m };
+            yield return new object[] { U.Day, U.Second, 1m, 86400m };
+            yield return new object[] { U.Kilogram, U.Ounce, 1m, 35.2740m };
+            yield return new object[] { U.Fahrenheit, U.Celsius, 100m, 37.7778m };
+            yield return new object[] { U.Fahrenheit, U.Kelvin, 100m, 310.9278m };
+            yield return new object[] { U.Pascal, U.Psi, 1m, 6894m };
+            yield return new object[] { U.Kilometer, U.Mile, 100m, 62.1371m };
         }
 
         public IEnumerable<object[]> ComplexConversionChecksSource()
         {
-            yield return new object[] { U.Kilometer / U.Hour, U.Mile / U.Hour, 100, 62.1371 };
-            yield return new object[] {U.Joule, U.Kilogram, 89875517873681764, 1};
-            yield return new object[] {U.Meter.Squared(), U.Hectare, 10000, 1};
-            yield return new object[] { U.Kilometer / U.Hour, U.Meter / U.Second, 100, 27.7778 };
+            yield return new object[] { U.Kilometer / U.Hour, U.Mile / U.Hour, 100m, 62.1371m };
+            yield return new object[] {U.Joule, U.Kilogram, 89875517873681764m, 1m};
+            yield return new object[] {U.Meter.Squared(), U.Hectare, 10000m, 1m};
+            yield return new object[] { U.Kilometer / U.Hour, U.Meter / U.Second, 100m, 27.7778m };
         }
 
         public IEnumerable<object[]> MeasureMultiplicationTestsSource()
@@ -38,7 +38,7 @@ namespace NMeasure.Tests
         }
 
         [Test, TestCaseSource("BasicConversionChecksSource")]
-        public void BasicConversionChecks(Unit from, Unit to, double input, double expectedOutput)
+        public void BasicConversionChecks(Unit from, Unit to, decimal input, decimal expectedOutput)
         {
             var m = (Measure) input*from;
             var m2 = m.ConvertTo(to);
@@ -46,7 +46,7 @@ namespace NMeasure.Tests
         }
 
         [Test, TestCaseSource("ComplexConversionChecksSource")]
-        public void ComplexConversionChecks(Unit from, Unit to, double input, double expectedOutput)
+        public void ComplexConversionChecks(Unit from, Unit to, decimal input, decimal expectedOutput)
         {
             var m = (Measure)input * from;
             var m2 = m.ConvertTo(to);
@@ -66,7 +66,7 @@ namespace NMeasure.Tests
             var m = 3 * U.Meter;
             var m2 = m - (10 * U.Inch);
             m2.Unit.IsEqualTo(U.Meter);
-            m2.Value.IsEqualTo(2.746);
+            m2.Value.IsEqualTo(2.746m);
